refactor: `LocationResponse` 를 `spi` 패키지로 이동
