+++ src/main/java/org/leisureup/travel/internal/travel/service/TravelService.java
@@ -1,18 +1,14 @@
 package org.leisureup.travel.internal.travel.service;
 
+import java.util.*;
 import lombok.*;
-import org.leisureup.global.exception.NotFound;
+import org.leisureup.global.exception.*;
 import org.leisureup.global.response.*;
-import org.leisureup.location.internal.dto.response.LocationResponse;
-import org.leisureup.location.spi.LocationQueryPort;
+import org.leisureup.location.spi.*;
 import org.leisureup.travel.internal.travel.domain.*;
 import org.leisureup.travel.internal.travel.dto.*;
 import org.leisureup.travel.internal.travel.repository.*;
 import org.springframework.stereotype.*;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
